price-reporter: Allow USDT/USDC stream
